Extend benchmark to report allocations
This is as a prelude to trying to reduce allocations.

Inital result for high level benchmark

Allocated total: 348802552
Allocated per:   1331

That is, 1331 bytes per IO operation. This includes allocations from the
benchmark itself. Indeed profiling indicates that most of the
allocations are from the benchmark itself rather than the library.