Adjust high level benchmark approach to reduce allocations
Previously we generated a large list of a random permutation of all the
4k disk blocks in the input file, and then broke that up into groups of
32 blocks. We used forConcurrently_ over the groups which starts one
thread per group.

This approach did excessive work and allocated a lot. For starters the
random permutation is expensive in time and space. It needed to track a
set of remaining choices. And it had to allocate the whole list because
the forConcurrently_ forces the list (of groups). And a thread per 32
blocks is a lot of threads once one uses large files. For our example
target file of 262144 blocks, this amounts to 8192 threads. Each one has
a stack, so it's a lot of allocations in total.

The new approach is:
1. Use random block indexes rather than a permutation. This leads to
   duplicates but is much cheaper because it needs little state. The
   best solution to the duplicates will be to use O_DIRECT to avoid
   caching effects.
2. Only use a small number of threads and have each one submit a series
   of batches of work (still of size 32).

This uses much less state. Then in addition, we use a lower level
approach to generatng the random block indexes to reduce heap
allocations.

The new allocation results for the high level benchmark are:

Allocated total: 57904976
Allocated per:   221

That is, 221 bytes per IO op. This is without any changes yet to the
actual library, just the benchmark.