Rewrite submitIO yet again.
Try to do two things:
1. reduce allocations and improve performance slightly
2. improve the sync and async exception safety of submitIO

Trying to achieve exception safety is non-trivial and needs masking and
multiple levels of exception handlers.

The one remaining difficulty is that if we get a sync exception from the
lower level submit operation then we do not yet reset the ring to clear
the SQEs that we filled out.