Use one kernel uring per RTS capability
The intention is to improve performance when using multiple
capabilities by reducing contention and submitting I/O from multiple
cores.

I/O is submitted to the ring for the capability that the thread is
currently running on. A completion thread runs per uring, and thus per
capability. It is safe if threads are migrated between submission and
completion.