Set the CQ ring size to be the concurrency limit
This is a workaround that prevents overflows of the CQ ring. This also
prevents the "uninterruptible-sleep process" problem from occurring,
which I was trying to debug in #27, though it is not clear what aspect
of overflows exactly is causing processes to go into uninterruptible
sleep.

One way of ensuring there is no CQ ring overflow is to make sure there
can never be more SQE's in flight than there is space for in the CQ
ring. The simplest way to achieve this is to set the CQ ring size to
the concurrency limit. The uring context is initialised with a batch
size and a concurrency limit, so these values are readily available.

Also adjust the low level benchmark to use the same default parameters.