Make the default concurrency limit twice the default batch size
It's a workaround that prevents overflows of the CQ ring.  This also prevents
the "uninterruptible-sleep process" problem from occurring, which I was trying
to debug in #27, though it is not clear what aspect of overflows exactly is
causing processes to go into uninterruptible sleep.

The uring context is initialised with a batch size and a concurrency limit. The
batch size is used as the size of the SQ ring, and the size of the CQ ring is
automatically set to twice the size of SQ ring. One way of ensuring there is no
CQ ring overflow is to make sure there can never be more SQE's in flight than
there is space for in the CQ ring. The simplest way to achieve this is to set
the concurrency limit to be twice the batch size.

The downside is that the high-level benchmark becomes a little bit slower.

Before:
```
Total I/O ops: 262144
Elapsed time:  1.459835906s
IOPS:          179571
```

After:
```
Total I/O ops: 262144
Elapsed time:  1.630825788s
IOPS:          160743
```